Defect Fixed

+++ src/main/java/com/rdsms/controller/BranchController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.rdsms.entity.Branch;
+import com.rdsms.entity.Candidate;
 import com.rdsms.service.IBranchService;
 
 @Controller
@@ -26,7 +27,7 @@ public class BranchController {
 	private IBranchService service;
 	
 	@GetMapping("branches")
-	@PreAuthorize("hasRole('USER')")
+	@PreAuthorize("hasRole('ADMIN')")
 	public ResponseEntity<List<Branch>> getBranches(){
 		List<Branch> branches = service.getBranches();
 		return new ResponseEntity<List<Branch>>(branches,HttpStatus.OK);
@@ -52,9 +53,14 @@ public class BranchController {
 	
 	@DeleteMapping("branches/{branchId}")
 	public ResponseEntity<String> deleteBranch(@PathVariable("branchId") int branchId) {
-		
-		boolean isDeleted = service.deleteBranch(branchId);
 		String message;
+		int candidates = service.getCandidatesByBranchId(branchId);
+		if(candidates>0) {
+			message = "This Branch is having "+ candidates + ", so it can not delete.";
+			return new ResponseEntity<String>(message,HttpStatus.BAD_REQUEST);
+		}
+		boolean isDeleted = service.deleteBranch(branchId);
+		
 		if(isDeleted) {
 			message = "Branch(BranchId="+branchId+") has been deleted successfully";
 		}else {
